refactor(client): simplify tree shaking analysis
