+++ packages/client/src/pages/TreeShaking/index.scss
@@ -9,16 +9,6 @@ $treePrefixCls: 'tree-shaking';
     flex-direction: column;
   }
 
-  .#{$treePrefixCls}-statement-side-effect {
-    text-decoration: underline;
-    background: #711f1c;
-  }
-
-  .#{$treePrefixCls}-statement-declaration-identifier {
-    text-decoration: underline;
-    background: rgb(44, 135, 236);
-  }
-
   .#{$treePrefixCls}-files-box {
     height: 700px;
   }
@@ -29,7 +19,6 @@ $treePrefixCls: 'tree-shaking';
   }
 
   .#{$treePrefixCls}-editor {
-    // margin-left: 40px
     flex-grow: 1;
     height: 700px;
 
@@ -44,12 +33,9 @@ $treePrefixCls: 'tree-shaking';
     height: calc(100% - 48px);
   }
 
-  .#{$treePrefixCls}-export-table {
+  .#{$treePrefixCls}-table {
     .ant-card-body {
       padding: 0;
-      .ant-divider {
-        margin: 24px;
-      }
     }
     .ant-table-wrapper .ant-table-tbody > tr:last-child > td {
       border-bottom: none;
@@ -63,31 +49,6 @@ $treePrefixCls: 'tree-shaking';
           border-start-end-radius: 0;
         }
       }
-    }
-  }
-
-  .#{$treePrefixCls}-side-effect-list {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    .#{$treePrefixCls}-side-effect-list-item {
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    }
-  }
-}
-.monaco-hover .markdown-hover.hover-row {
-  hr {
-    margin: 2px;
-    background-color: gray;
-  }
-
-  div[data-code='tree-shaking-hover'] ol {
-    margin: 4px 0 10px 0;
-    padding-left: 4px;
-    pre {
-      margin: 2px 0;
     }
   }
 }
