+++ packages/client/src/pages/TreeShaking/index.tsx
@@ -1,99 +1,36 @@
-import React, { useMemo, useState } from 'react';
-import { Card, Col, Row, Radio } from 'antd';
-import { Lodash } from '@rsdoctor/shared/common-browser';
-import { FileSearchOutlined } from '@ant-design/icons';
 import { SDK } from '@rsdoctor/shared/types';
+import { Card, Col, Empty, Row, Table, Typography } from 'antd';
+import React, { useState } from 'react';
 import path from 'path-browserify';
-import { ServerAPIProvider, withManifestData } from 'src/components/Manifest';
-import { Size } from '../../constants';
+import { FileSearchOutlined } from '@ant-design/icons';
 import { FileTree } from '../../components/FileTree';
 import { KeywordInput } from '../../components/Form/keyword';
-import { getTreeFilesDefaultExpandedKeys, useFileStructures } from './utils';
-import {
-  fetchManifest,
-  useModuleGraphInstanceByModuleGraph,
-} from '../../utils';
-import { CodeEditor } from './editor';
-import { TreeShakingDataBoundary } from './data-boundary';
-import { SideEffectsAnalysis } from './side-effects';
-import { TreeShakingTable } from './table';
-import type { TableKind, SetEditorStatus } from './types';
+import { ServerAPIProvider, withManifestData } from '../../components/Manifest';
+import { fetchManifest } from '../../utils';
+import { CodeViewer } from '../../components/base';
+import { Size } from '../../constants';
+import { getSideEffectModules, getSideEffectTree } from './side-effects-data';
 
 import './index.scss';
 export * from './constants';
 
-const Component: React.FC<{
+function Component({
+  data,
+  treeShaking,
+  cwd,
+}: {
   data: SDK.ModuleGraphData;
-  cwd: string;
   treeShaking?: SDK.TreeShakingData;
-}> = ({ data, cwd, treeShaking }) => {
-  const moduleGraph = useModuleGraphInstanceByModuleGraph(data);
-
-  if (treeShaking) {
-    return (
-      <SideEffectsAnalysis
-        modules={data.modules}
-        data={treeShaking}
-        cwd={cwd}
-      />
-    );
-  }
-
-  return (
-    <TreeShakingDataBoundary moduleGraph={moduleGraph}>
-      <Analysis moduleGraph={moduleGraph} cwd={cwd} />
-    </TreeShakingDataBoundary>
-  );
-};
-
-const Analysis: React.FC<{
-  moduleGraph: SDK.ModuleGraphInstance;
   cwd: string;
-}> = ({ moduleGraph, cwd }) => {
-  const [searchInput, setSearchInput] = useState('');
-  const [toLine, setToLine] = useState(1);
-  const [ranges, setRanges] = useState<SDK.SourceRange[]>([]);
-  const [tableKind, setTableKind] = useState<TableKind>('side-effect');
-  const filteredModules = useMemo(
-    () =>
-      moduleGraph
-        .getModules()
-        .filter(
-          (item) =>
-            item.kind === SDK.ModuleKind.Normal &&
-            path.basename(item.path).includes(searchInput),
-        ),
-    [searchInput],
-  );
-
-  const entry = useMemo(
-    () => filteredModules.find((item) => item.isEntry)!,
-    [],
-  );
-  const [selectedModule, setSelectedModule] = useState(
-    entry || (filteredModules?.length && filteredModules?.[0]),
-  );
-  const files = useFileStructures(
-    filteredModules,
-    moduleGraph,
-    searchInput,
-    selectedModule,
-    (file) => {
-      setEditorData(
-        filteredModules.find((item) => item.path === file)!,
-        [],
-      );
-    },
-    cwd,
-  );
-  const setEditorData: SetEditorStatus = (module, ranges, line) => {
-    setSelectedModule(module);
-    setRanges(ranges.slice());
-
-    if (Lodash.isNumber(line)) {
-      setToLine(line);
-    }
-  };
+}): React.ReactElement {
+  const [search, setSearch] = useState('');
+  const [selectedId, setSelectedId] = useState<number>();
+  const [codeIndex, setCodeIndex] = useState(0);
+  const filtered = getSideEffectModules(data.modules, treeShaking, search);
+  const selected =
+    filtered.find((module) => module.id === selectedId) ?? filtered[0];
+  const codes = selected ? treeShaking!.sideEffectCodes[selected.id] : [];
+  const code = codes[codeIndex] ?? codes[0];
 
   return (
     <Card
@@ -104,76 +41,141 @@ const Analysis: React.FC<{
       <Row
         justify="space-between"
         align="middle"
+        gutter={[Size.BasePadding, Size.BasePadding]}
         style={{ marginBottom: Size.BasePadding, marginTop: Size.BasePadding }}
       >
-        <KeywordInput
-          icon={<FileSearchOutlined />}
-          width={400}
-          label="FileName"
-          placeholder="search filename by keyword"
-          style={{ width: 'auto' }}
-          onChange={(e) => setSearchInput(e)}
-        />
-        <Radio.Group
-          value={tableKind}
-          onChange={({ target }) => setTableKind(target.value)}
-        >
-          <Radio.Button value="side-effect">SideEffects</Radio.Button>
-          <Radio.Button value="export">Export Variables</Radio.Button>
-        </Radio.Group>
-      </Row>
-      <Row
-        justify="space-between"
-        align="top"
-        wrap={false}
-        gutter={[Size.BasePadding, Size.BasePadding]}
-      >
-        <Col span={7}>
-          <Card
-            title={`Total Files: ${filteredModules.length}`}
-            className="tree-shaking-files-box"
-          >
-            <FileTree
-              style={{ height: '80%' }}
-              className="tree-shaking-files"
-              selectedKeys={selectedModule ? [selectedModule.path] : []}
-              defaultExpandedKeys={getTreeFilesDefaultExpandedKeys(files)}
-              treeData={files}
-              expandAction="click"
-            />
-          </Card>
-        </Col>
-        <Col span={9}>
-          <ServerAPIProvider
-            api={SDK.ServerAPI.API.GetModuleCodeByModuleId}
-            body={{ moduleId: selectedModule.id }}
-          >
-            {(source) => {
-              return (
-                <CodeEditor
-                  module={selectedModule}
-                  moduleGraph={moduleGraph}
-                  ranges={ranges}
-                  toLine={toLine}
-                  setEditorData={setEditorData}
-                  source={source}
-                />
-              );
+        <Col>
+          <KeywordInput
+            icon={<FileSearchOutlined />}
+            width={400}
+            label="FileName"
+            placeholder="search filename by keyword"
+            style={{ width: 'auto' }}
+            onChange={(value) => {
+              setSearch(value);
+              setSelectedId(undefined);
+              setCodeIndex(0);
             }}
-          </ServerAPIProvider>
-        </Col>
-        <Col flex={1}>
-          <TreeShakingTable
-            kind={tableKind}
-            setEditorData={setEditorData}
-            module={selectedModule}
-            moduleGraph={moduleGraph}
           />
         </Col>
       </Row>
+      {selected ? (
+        <Row
+          align="top"
+          wrap={false}
+          gutter={[Size.BasePadding, Size.BasePadding]}
+        >
+          <Col span={7} style={{ minWidth: 0 }}>
+            <Card
+              title={`Total Files: ${filtered.length}`}
+              className="tree-shaking-files-box"
+            >
+              <FileTree
+                key={search}
+                className="tree-shaking-files"
+                treeData={getSideEffectTree(filtered, cwd)}
+                defaultExpandAll
+                selectedKeys={[`module:${selected.id}`]}
+                onSelect={(keys) => {
+                  const key = String(keys[0] ?? '');
+                  if (key.startsWith('module:')) {
+                    setSelectedId(Number(key.slice(7)));
+                    setCodeIndex(0);
+                  }
+                }}
+              />
+            </Card>
+          </Col>
+          <Col span={9} style={{ minWidth: 0 }}>
+            <Card
+              title={
+                <Typography.Text ellipsis={{ tooltip: selected.path }}>
+                  {path.basename(selected.path)}
+                </Typography.Text>
+              }
+              className="tree-shaking-editor"
+            >
+              <ServerAPIProvider
+                key={selected.id}
+                api={SDK.ServerAPI.API.GetModuleCodeByModuleId}
+                body={{ moduleId: selected.id }}
+              >
+                {(source) =>
+                  source.source || source.parsedSource ? (
+                    <CodeViewer
+                      key={`${selected.id}:${codeIndex}`}
+                      code={source.source || source.parsedSource}
+                      filePath={selected.path}
+                      defaultLine={code?.startLine}
+                      ranges={
+                        code
+                          ? [{ start: { line: code.startLine, column: 0 } }]
+                          : []
+                      }
+                      style={{ height: 640, minWidth: 0 }}
+                    />
+                  ) : (
+                    <Empty description="Source code is not included in this report." />
+                  )
+                }
+              </ServerAPIProvider>
+            </Card>
+          </Col>
+          <Col span={8} style={{ minWidth: 0 }}>
+            <Card
+              title={`SideEffects: ${codes.length}`}
+              className="tree-shaking-table"
+            >
+              <Table
+                size="small"
+                pagination={false}
+                scroll={{ y: 640 }}
+                rowKey={(_, index) => String(index)}
+                dataSource={codes}
+                columns={[
+                  { title: 'Line', dataIndex: 'startLine', width: 64 },
+                  {
+                    title: 'Code',
+                    dataIndex: 'code',
+                    render: (text: string) => (
+                      <pre
+                        style={{
+                          whiteSpace: 'pre-wrap',
+                          overflowWrap: 'anywhere',
+                          margin: 0,
+                        }}
+                      >
+                        {text}
+                      </pre>
+                    ),
+                  },
+                ]}
+                onRow={(_, index) => ({
+                  onClick: () => setCodeIndex(index ?? 0),
+                  style: {
+                    cursor: 'pointer',
+                    background:
+                      index === codeIndex
+                        ? 'var(--ant-color-primary-bg)'
+                        : undefined,
+                  },
+                })}
+              />
+            </Card>
+          </Col>
+        </Row>
+      ) : (
+        <Empty
+          description={
+            search
+              ? 'No matching files'
+              : 'No side-effect code was recorded for modules in this report.'
+          }
+        />
+      )}
     </Card>
   );
-};
+}
 
 export const TreeShakingPage = withManifestData(
   fetchManifest,
